+++ CHANGES.rst
@@ -1,3 +1,11 @@
+[unreleased]
+------------
+
+Added
+^^^^^
+- SCIM ``attributes`` and ``excludedAttributes`` query parameter support.
+- SCIM ``POST /.search`` endpoint.
+
 [0.2.3] - 2026-03-24
 --------------------